+++ CHANGELOG.md
@@ -13,6 +13,7 @@ x.x.x - TODO
   * ...
 
 3.x.x - 2024xxxx
+  * Move base url and slugify from Route to UriGenerator
   * Use RequestContext and instance methods in Framework for future adapters
   * Add HasRouteTrait to simplify handler route() calls and future migrations
   * Split UriGenerator from Route + add HandlerManager for framework adapters